requirements.txt
Flask==3.1.0
flask-cors==5.0.0
flask-jwt-extended==4.7.1
python-dotenv
stripe
google-genai
Flask-Limiter
supabase